pyproject.toml
[project]
name = "demo-uv-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow>=4.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
